chore: cleanup migration after 3.7.0 release
