+++ crates/contract/src/lib.rs
@@ -11,7 +11,7 @@ pub mod tee;
 pub mod update;
 #[cfg(feature = "dev-utils")]
 pub mod utils;
-pub mod v3_6_0_state;
+pub mod v3_7_0_state;
 
 #[cfg(feature = "bench-contract-methods")]
 mod bench;
@@ -1673,11 +1673,11 @@ impl MpcContract {
     pub fn migrate() -> Result<Self, Error> {
         log!("migrating contract");
 
-        match try_state_read::<v3_6_0_state::MpcContract>() {
+        match try_state_read::<v3_7_0_state::MpcContract>() {
             Ok(Some(state)) => return Ok(state.into()),
             Ok(None) => return Err(InvalidState::ContractStateIsMissing.into()),
             Err(err) => {
-                log!("failed to deserialize state into v3.6.0 state: {:?}", err);
+                log!("failed to deserialize state into v3.7.0 state: {:?}", err);
             }
         };
