+++ crates/contract/src/v3_7_0_state.rs
@@ -1,0 +1,67 @@
+//! ## Overview
+//! This module stores the previous contract state—the one you want to migrate from.
+//! The goal is to describe the data layout _exactly_ as it existed before.
+//!
+//! ## Guideline
+//! In theory, you could copy-paste every struct from the specific commit you're migrating from.
+//! However, this approach (a) requires manual effort from a developer and (b) increases the binary size.
+//! A better approach: only copy the structures that have changed and import the rest from the existing codebase.
+
+use borsh::{BorshDeserialize, BorshSerialize};
+use near_sdk::env;
+use near_sdk::store::LookupMap;
+
+use crate::{
+    node_migrations::NodeMigrations,
+    primitives::{
+        ckd::CKDRequest,
+        signature::{SignatureRequest, YieldIndex},
+    },
+    state::ProtocolContractState,
+    tee::tee_state::TeeState,
+    update::ProposedUpdates,
+    Config, ForeignChainPolicyVotes, StaleData,
+};
+use near_mpc_contract_interface::types as dtos;
+
+#[derive(Debug, BorshSerialize, BorshDeserialize)]
+pub struct MpcContract {
+    protocol_state: ProtocolContractState,
+    pending_signature_requests: LookupMap<SignatureRequest, YieldIndex>,
+    pending_ckd_requests: LookupMap<CKDRequest, YieldIndex>,
+    pending_verify_foreign_tx_requests:
+        LookupMap<dtos::VerifyForeignTransactionRequest, YieldIndex>,
+    proposed_updates: ProposedUpdates,
+    foreign_chain_policy: dtos::ForeignChainPolicy,
+    foreign_chain_policy_votes: ForeignChainPolicyVotes,
+    config: Config,
+    tee_state: TeeState,
+    accept_requests: bool,
+    node_migrations: NodeMigrations,
+    stale_data: StaleData,
+    metrics: dtos::Metrics,
+}
+
+impl From<MpcContract> for crate::MpcContract {
+    fn from(value: MpcContract) -> Self {
+        let crate::ProtocolContractState::Running(_running_state) = &value.protocol_state else {
+            env::panic_str("Contract must be in running state when migrating.");
+        };
+
+        Self {
+            protocol_state: value.protocol_state,
+            pending_signature_requests: value.pending_signature_requests,
+            pending_ckd_requests: value.pending_ckd_requests,
+            pending_verify_foreign_tx_requests: value.pending_verify_foreign_tx_requests,
+            proposed_updates: value.proposed_updates,
+            foreign_chain_policy: value.foreign_chain_policy,
+            foreign_chain_policy_votes: value.foreign_chain_policy_votes,
+            config: value.config,
+            tee_state: value.tee_state,
+            accept_requests: value.accept_requests,
+            node_migrations: value.node_migrations,
+            stale_data: crate::StaleData {},
+            metrics: value.metrics,
+        }
+    }
+}
